+++ packages/core/src/managed/graphql/graphql-sdk.ts
@@ -55,6 +55,8 @@ export interface MdArtifact {
   reference: Scalars['String'];
   versions?: Maybe<Array<MdArtifactVersionInEnvironment>>;
   pinnedVersion?: Maybe<MdPinnedVersion>;
+  latestApprovedVersion?: Maybe<MdArtifactVersionInEnvironment>;
+  resources?: Maybe<Array<MdResource>>;
 }
 
 export interface MdArtifactVersionsArgs {
@@ -87,7 +89,6 @@ export interface MdArtifactVersionInEnvironment {
   buildNumber?: Maybe<Scalars['String']>;
   createdAt?: Maybe<Scalars['InstantTime']>;
   deployedAt?: Maybe<Scalars['InstantTime']>;
-  resources?: Maybe<Array<MdResource>>;
   gitMetadata?: Maybe<MdGitMetadata>;
   packageDiff?: Maybe<MdPackageDiff>;
   environment?: Maybe<Scalars['String']>;
@@ -120,6 +121,7 @@ export interface MdConfig {
   updatedAt?: Maybe<Scalars['InstantTime']>;
   rawConfig?: Maybe<Scalars['String']>;
   processedConfig?: Maybe<Scalars['String']>;
+  previewEnvironmentsConfigured?: Maybe<Scalars['Boolean']>;
 }
 
 export interface MdConstraint {
@@ -142,6 +144,20 @@ export interface MdConstraintStatusPayload {
   version: Scalars['String'];
   reference: Scalars['String'];
   status: MdConstraintStatus;
+}
+
+export interface MdDeployLocation {
+  __typename?: 'MdDeployLocation';
+  account?: Maybe<Scalars['String']>;
+  region?: Maybe<Scalars['String']>;
+  sublocations?: Maybe<Array<Scalars['String']>>;
+}
+
+export interface MdDeployTarget {
+  __typename?: 'MdDeployTarget';
+  cloudProvider?: Maybe<Scalars['String']>;
+  location?: Maybe<MdDeployLocation>;
+  status?: Maybe<MdRolloutTargetStatus>;
 }
 
 export interface MdDismissNotificationPayload {
@@ -168,6 +184,16 @@ export interface MdEnvironmentState {
 }
 
 export type MdEventLevel = 'SUCCESS' | 'INFO' | 'WARNING' | 'ERROR';
+
+export interface MdExecutionSummary {
+  __typename?: 'MdExecutionSummary';
+  status: MdTaskStatus;
+  currentStage?: Maybe<MdStageDetail>;
+  stages?: Maybe<Array<MdStageDetail>>;
+  deployTargets?: Maybe<Array<MdDeployTarget>>;
+  completedDeployTargets?: Maybe<Array<MdDeployTarget>>;
+  error?: Maybe<Scalars['String']>;
+}
 
 export interface MdGitIntegration {
   __typename?: 'MdGitIntegration';
@@ -305,6 +331,7 @@ export interface MdResource {
 
 export interface MdResourceActuationState {
   __typename?: 'MdResourceActuationState';
+  resourceId: Scalars['String'];
   status: MdResourceActuationStatus;
   reason?: Maybe<Scalars['String']>;
   event?: Maybe<Scalars['String']>;
@@ -317,6 +344,16 @@ export interface MdResourceTask {
   __typename?: 'MdResourceTask';
   id: Scalars['String'];
   name: Scalars['String'];
+  running: Scalars['Boolean'];
+  summary?: Maybe<MdExecutionSummary>;
+}
+
+export interface MdRestartConstraintEvaluationPayload {
+  application: Scalars['String'];
+  environment: Scalars['String'];
+  type: Scalars['String'];
+  reference: Scalars['String'];
+  version: Scalars['String'];
 }
 
 export interface MdRetryArtifactActionPayload {
@@ -327,6 +364,34 @@ export interface MdRetryArtifactActionPayload {
   actionId: Scalars['String'];
   actionType: MdActionType;
 }
+
+export type MdRolloutTargetStatus = 'NOT_STARTED' | 'RUNNING' | 'SUCCEEDED' | 'FAILED';
+
+export interface MdStageDetail {
+  __typename?: 'MdStageDetail';
+  id?: Maybe<Scalars['String']>;
+  type?: Maybe<Scalars['String']>;
+  name?: Maybe<Scalars['String']>;
+  startTime?: Maybe<Scalars['InstantTime']>;
+  endTime?: Maybe<Scalars['InstantTime']>;
+  status?: Maybe<MdTaskStatus>;
+  refId?: Maybe<Scalars['String']>;
+  requisiteStageRefIds?: Maybe<Array<Scalars['String']>>;
+}
+
+export type MdTaskStatus =
+  | 'NOT_STARTED'
+  | 'RUNNING'
+  | 'PAUSED'
+  | 'SUSPENDED'
+  | 'SUCCEEDED'
+  | 'FAILED_CONTINUE'
+  | 'TERMINAL'
+  | 'CANCELED'
+  | 'REDIRECT'
+  | 'STOPPED'
+  | 'BUFFERED'
+  | 'SKIPPED';
 
 export interface MdToggleResourceManagementPayload {
   id: Scalars['ID'];
@@ -355,6 +420,7 @@ export interface MdVersionVeto {
 export interface Mutation {
   __typename?: 'Mutation';
   updateConstraintStatus?: Maybe<Scalars['Boolean']>;
+  restartConstraintEvaluation?: Maybe<Scalars['Boolean']>;
   toggleManagement?: Maybe<Scalars['Boolean']>;
   pinArtifactVersion?: Maybe<Scalars['Boolean']>;
   markArtifactVersionAsBad?: Maybe<Scalars['Boolean']>;
@@ -369,6 +435,10 @@ export interface Mutation {
 
 export interface MutationUpdateConstraintStatusArgs {
   payload: MdConstraintStatusPayload;
+}
+
+export interface MutationRestartConstraintEvaluationArgs {
+  payload: MdRestartConstraintEvaluationPayload;
 }
 
 export interface MutationToggleManagementArgs {
@@ -491,7 +561,8 @@ export type FetchApplicationQueryVariables = Exact<{
 
 export type FetchApplicationQuery = { __typename?: 'Query' } & {
   application?: Maybe<
-    { __typename?: 'MdApplication' } & Pick<MdApplication, 'id' | 'name' | 'account'> & {
+    { __typename?: 'MdApplication' } & Pick<MdApplication, 'id' | 'name'> & {
+        config?: Maybe<{ __typename?: 'MdConfig' } & Pick<MdConfig, 'id' | 'previewEnvironmentsConfigured'>>;
         environments: Array<
           { __typename?: 'MdEnvironment' } & Pick<MdEnvironment, 'isDeleting'> & {
               state: { __typename?: 'MdEnvironmentState' } & Pick<MdEnvironmentState, 'id'> & {
@@ -525,6 +596,45 @@ export type FetchApplicationQuery = { __typename?: 'Query' } & {
   >;
 };
 
+export type FetchCurrentVersionQueryVariables = Exact<{
+  appName: Scalars['String'];
+}>;
+
+export type FetchCurrentVersionQuery = { __typename?: 'Query' } & {
+  application?: Maybe<
+    { __typename?: 'MdApplication' } & Pick<MdApplication, 'id' | 'name'> & {
+        environments: Array<
+          { __typename?: 'MdEnvironment' } & Pick<MdEnvironment, 'id' | 'name'> & {
+              state: { __typename?: 'MdEnvironmentState' } & {
+                artifacts?: Maybe<
+                  Array<
+                    { __typename?: 'MdArtifact' } & Pick<MdArtifact, 'id' | 'name' | 'reference' | 'environment'> & {
+                        versions?: Maybe<
+                          Array<
+                            { __typename?: 'MdArtifactVersionInEnvironment' } & Pick<
+                              MdArtifactVersionInEnvironment,
+                              'id' | 'version' | 'buildNumber' | 'createdAt'
+                            > & {
+                                gitMetadata?: Maybe<
+                                  { __typename?: 'MdGitMetadata' } & Pick<MdGitMetadata, 'commit'> & {
+                                      commitInfo?: Maybe<
+                                        { __typename?: 'MdCommitInfo' } & Pick<MdCommitInfo, 'sha' | 'message'>
+                                      >;
+                                    }
+                                >;
+                              }
+                          >
+                        >;
+                      }
+                  >
+                >;
+              };
+            }
+        >;
+      }
+  >;
+};
+
 export type FetchVersionsHistoryQueryVariables = Exact<{
   appName: Scalars['String'];
   limit?: Maybe<Scalars['Int']>;
@@ -532,7 +642,7 @@ export type FetchVersionsHistoryQueryVariables = Exact<{
 
 export type FetchVersionsHistoryQuery = { __typename?: 'Query' } & {
   application?: Maybe<
-    { __typename?: 'MdApplication' } & Pick<MdApplication, 'id' | 'name' | 'account'> & {
+    { __typename?: 'MdApplication' } & Pick<MdApplication, 'id' | 'name'> & {
         environments: Array<
           { __typename?: 'MdEnvironment' } & {
             state: { __typename?: 'MdEnvironmentState' } & Pick<MdEnvironmentState, 'id'> & {
@@ -695,7 +805,7 @@ export type FetchApplicationManagementDataQuery = { __typename?: 'Query' } & {
         gitIntegration?: Maybe<
           { __typename?: 'MdGitIntegration' } & Pick<
             MdGitIntegration,
-            'id' | 'repository' | 'branch' | 'isEnabled' | 'link'
+            'id' | 'repository' | 'branch' | 'isEnabled' | 'link' | 'manifestPath'
           >
         >;
       }
@@ -774,6 +884,21 @@ export type ImportDeliveryConfigMutationVariables = Exact<{
 }>;
 
 export type ImportDeliveryConfigMutation = { __typename?: 'Mutation' } & Pick<Mutation, 'importDeliveryConfig'>;
+
+export type ToggleResourceManagementMutationVariables = Exact<{
+  payload?: Maybe<MdToggleResourceManagementPayload>;
+}>;
+
+export type ToggleResourceManagementMutation = { __typename?: 'Mutation' } & Pick<Mutation, 'toggleResourceManagement'>;
+
+export type RestartConstraintEvaluationMutationVariables = Exact<{
+  payload: MdRestartConstraintEvaluationPayload;
+}>;
+
+export type RestartConstraintEvaluationMutation = { __typename?: 'Mutation' } & Pick<
+  Mutation,
+  'restartConstraintEvaluation'
+>;
 
 export const ActionDetailsFragmentDoc = gql`
   fragment actionDetails on MdAction {
@@ -877,7 +1002,10 @@ export const FetchApplicationDocument = gql`
     application(appName: $appName) {
       id
       name
-      account
+      config {
+        id
+        previewEnvironmentsConfigured
+      }
       environments {
         ...baseEnvironmentFields
         isDeleting
@@ -950,12 +1078,85 @@ export function useFetchApplicationLazyQuery(
 export type FetchApplicationQueryHookResult = ReturnType<typeof useFetchApplicationQuery>;
 export type FetchApplicationLazyQueryHookResult = ReturnType<typeof useFetchApplicationLazyQuery>;
 export type FetchApplicationQueryResult = Apollo.QueryResult<FetchApplicationQuery, FetchApplicationQueryVariables>;
+export const FetchCurrentVersionDocument = gql`
+  query fetchCurrentVersion($appName: String!) {
+    application(appName: $appName) {
+      id
+      name
+      environments {
+        id
+        name
+        state {
+          artifacts {
+            id
+            name
+            reference
+            environment
+            versions(statuses: [CURRENT]) {
+              id
+              version
+              buildNumber
+              createdAt
+              gitMetadata {
+                commit
+                commitInfo {
+                  sha
+                  message
+                }
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+`;
+
+/**
+ * __useFetchCurrentVersionQuery__
+ *
+ * To run a query within a React component, call `useFetchCurrentVersionQuery` and pass it any options that fit your needs.
+ * When your component renders, `useFetchCurrentVersionQuery` returns an object from Apollo Client that contains loading, error, and data properties
+ * you can use to render your UI.
+ *
+ * @param baseOptions options that will be passed into the query, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options;
+ *
+ * @example
+ * const { data, loading, error } = useFetchCurrentVersionQuery({
+ *   variables: {
+ *      appName: // value for 'appName'
+ *   },
+ * });
+ */
+export function useFetchCurrentVersionQuery(
+  baseOptions: Apollo.QueryHookOptions<FetchCurrentVersionQuery, FetchCurrentVersionQueryVariables>,
+) {
+  const options = { ...defaultOptions, ...baseOptions };
+  return Apollo.useQuery<FetchCurrentVersionQuery, FetchCurrentVersionQueryVariables>(
+    FetchCurrentVersionDocument,
+    options,
+  );
+}
+export function useFetchCurrentVersionLazyQuery(
+  baseOptions?: Apollo.LazyQueryHookOptions<FetchCurrentVersionQuery, FetchCurrentVersionQueryVariables>,
+) {
+  const options = { ...defaultOptions, ...baseOptions };
+  return Apollo.useLazyQuery<FetchCurrentVersionQuery, FetchCurrentVersionQueryVariables>(
+    FetchCurrentVersionDocument,
+    options,
+  );
+}
+export type FetchCurrentVersionQueryHookResult = ReturnType<typeof useFetchCurrentVersionQuery>;
+export type FetchCurrentVersionLazyQueryHookResult = ReturnType<typeof useFetchCurrentVersionLazyQuery>;
+export type FetchCurrentVersionQueryResult = Apollo.QueryResult<
+  FetchCurrentVersionQuery,
+  FetchCurrentVersionQueryVariables
+>;
 export const FetchVersionsHistoryDocument = gql`
   query fetchVersionsHistory($appName: String!, $limit: Int) {
     application(appName: $appName) {
       id
       name
-      account
       environments {
         ...baseEnvironmentFields
         state {
@@ -1309,6 +1510,7 @@ export const FetchApplicationManagementDataDocument = gql`
         branch
         isEnabled
         link
+        manifestPath
       }
     }
   }
@@ -1833,4 +2035,91 @@ export type ImportDeliveryConfigMutationResult = Apollo.MutationResult<ImportDel
 export type ImportDeliveryConfigMutationOptions = Apollo.BaseMutationOptions<
   ImportDeliveryConfigMutation,
   ImportDeliveryConfigMutationVariables
+>;
+export const ToggleResourceManagementDocument = gql`
+  mutation ToggleResourceManagement($payload: MdToggleResourceManagementPayload) {
+    toggleResourceManagement(payload: $payload)
+  }
+`;
+export type ToggleResourceManagementMutationFn = Apollo.MutationFunction<
+  ToggleResourceManagementMutation,
+  ToggleResourceManagementMutationVariables
+>;
+
+/**
+ * __useToggleResourceManagementMutation__
+ *
+ * To run a mutation, you first call `useToggleResourceManagementMutation` within a React component and pass it any options that fit your needs.
+ * When your component renders, `useToggleResourceManagementMutation` returns a tuple that includes:
+ * - A mutate function that you can call at any time to execute the mutation
+ * - An object with fields that represent the current status of the mutation's execution
+ *
+ * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
+ *
+ * @example
+ * const [toggleResourceManagementMutation, { data, loading, error }] = useToggleResourceManagementMutation({
+ *   variables: {
+ *      payload: // value for 'payload'
+ *   },
+ * });
+ */
+export function useToggleResourceManagementMutation(
+  baseOptions?: Apollo.MutationHookOptions<ToggleResourceManagementMutation, ToggleResourceManagementMutationVariables>,
+) {
+  const options = { ...defaultOptions, ...baseOptions };
+  return Apollo.useMutation<ToggleResourceManagementMutation, ToggleResourceManagementMutationVariables>(
+    ToggleResourceManagementDocument,
+    options,
+  );
+}
+export type ToggleResourceManagementMutationHookResult = ReturnType<typeof useToggleResourceManagementMutation>;
+export type ToggleResourceManagementMutationResult = Apollo.MutationResult<ToggleResourceManagementMutation>;
+export type ToggleResourceManagementMutationOptions = Apollo.BaseMutationOptions<
+  ToggleResourceManagementMutation,
+  ToggleResourceManagementMutationVariables
+>;
+export const RestartConstraintEvaluationDocument = gql`
+  mutation RestartConstraintEvaluation($payload: MdRestartConstraintEvaluationPayload!) {
+    restartConstraintEvaluation(payload: $payload)
+  }
+`;
+export type RestartConstraintEvaluationMutationFn = Apollo.MutationFunction<
+  RestartConstraintEvaluationMutation,
+  RestartConstraintEvaluationMutationVariables
+>;
+
+/**
+ * __useRestartConstraintEvaluationMutation__
+ *
+ * To run a mutation, you first call `useRestartConstraintEvaluationMutation` within a React component and pass it any options that fit your needs.
+ * When your component renders, `useRestartConstraintEvaluationMutation` returns a tuple that includes:
+ * - A mutate function that you can call at any time to execute the mutation
+ * - An object with fields that represent the current status of the mutation's execution
+ *
+ * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
+ *
+ * @example
+ * const [restartConstraintEvaluationMutation, { data, loading, error }] = useRestartConstraintEvaluationMutation({
+ *   variables: {
+ *      payload: // value for 'payload'
+ *   },
+ * });
+ */
+export function useRestartConstraintEvaluationMutation(
+  baseOptions?: Apollo.MutationHookOptions<
+    RestartConstraintEvaluationMutation,
+    RestartConstraintEvaluationMutationVariables
+  >,
+) {
+  const options = { ...defaultOptions, ...baseOptions };
+  return Apollo.useMutation<RestartConstraintEvaluationMutation, RestartConstraintEvaluationMutationVariables>(
+    RestartConstraintEvaluationDocument,
+    options,
+  );
+}
+export type RestartConstraintEvaluationMutationHookResult = ReturnType<typeof useRestartConstraintEvaluationMutation>;
+export type RestartConstraintEvaluationMutationResult = Apollo.MutationResult<RestartConstraintEvaluationMutation>;
+export type RestartConstraintEvaluationMutationOptions = Apollo.BaseMutationOptions<
+  RestartConstraintEvaluationMutation,
+  RestartConstraintEvaluationMutationVariables
 >;